CXP-3032: Race condition during `local.buffer.dir` creation
When multiple tasks are starte on the same worker all of them are
trying to check if the directory exists and create if not. Sometimes
some of the tasks fail because the directory gets created after
the existence check and before the creation attempt by other tasks.

The fix is to do a few attempts of checking and creating the directory.